feat: 채팅 정보가 담긴 아코디언 유연하게 변경
